services/image-gen/requirements.txt
Flask==3.0.3
Pillow==10.3.0
pycairo==1.26.0
Werkzeug==3.0.3
minio==7.2.7
pytest==8.2.0
pytest-cov==5.0.0
